fix: refactor provider, add mcp env propagation
